mqtt-client/requirements.txt
# MQTT Client Dependencies
paho-mqtt==1.6.1
python-dotenv==1.0.0
requests==2.32.3